+++ lib/src/extensions/every_date_validator_list_ext.dart
@@ -1,0 +1,24 @@
+import '../everies/everies.dart';
+
+/// Extension methods related to [EveryDateValidator]s. Simply wrappers around
+/// creating [EveryDateValidatorIntersection], [EveryDateValidatorUnion] or
+/// [EveryDateValidatorDifference].
+extension EveryDateValidatorListExt on List<EveryDateValidator> {
+  /// Returns a [EveryDateValidatorIntersection] of this [List] of
+  /// [EveryDateValidator]s.
+  EveryDateValidatorIntersection get intersection {
+    return EveryDateValidatorIntersection(this);
+  }
+
+  /// Returns a [EveryDateValidatorDifference] of this [List] of
+  /// [EveryDateValidator]s.
+  EveryDateValidatorDifference get difference {
+    return EveryDateValidatorDifference(this);
+  }
+
+  /// Returns a [EveryDateValidatorUnion] of this [List] of
+  /// [EveryDateValidator]s.
+  EveryDateValidatorUnion get union {
+    return EveryDateValidatorUnion(this);
+  }
+}
