+++ test/extensions/day_in_year_test.dart
@@ -1,0 +1,16 @@
+import 'package:due_date/due_date.dart';
+import 'package:test/test.dart';
+
+void main () {
+  group('DayInYear on DateTime', () {
+    test('January 1st', () {
+      expect(DateTime(2022).dayInYear, 1);
+    });
+    test('December 31st 2022', () {
+      expect(DateTime(2022, DateTime.december, 31).dayInYear, 365);
+    });
+    test('December 31st 2020', () {
+      expect(DateTime(2020, DateTime.december, 31).dayInYear, 366);
+    });
+  });
+}
